Major update after census query!
-- all files cleaned to place roughly
-- moduleized code for query all data from air quality and census data base
-- json files created to record the query process and definition of the columns
