+++ .gitignore
@@ -1,2 +1,3 @@
 myEmailKey.txt
 *.csv
+censusapikey.txt
