+++ Concept_test_and_litreview/Meeting 19 May 2023.txt
@@ -1,0 +1,12 @@
+Things we need to do in the future:
+
+- Louise might look into additional variables from datasets in addition to ACS5.
+- Tian might look into creating a GUI to allow users to view different visualizations based on race, air quality, etc.
+- Tian is thinking about looking into how we are going to assign counties to each air quality station.
+- We are probably going to average the air quality data across all stations per county to produce a final figure.
+- SQL database because the air quality data is way too large.
+- We are going to have to think about how we are going to clean up the air quality data frame.
+- Air quality data: we want to get daily data on pollutants and condense that into a yearly measure by seeing how many days was there bad air quality by county.
+
+- Louise will look into selecting variables from race, ethnicity, population, income (possibly different types; poverty levels might be interesting), employment levels, commute lifestyle, industry, education levels.
+- Matthew will look into how to visualize based on counties only in the state.
