+++ US-Census-Data/CENSUS_API_TOOL/json/variables.json
@@ -1,0 +1,11 @@
+{
+    "COMMUTING TO WORK": ["DP03_0018E", "DP03_0018PE", "DP03_0019E", "DP03_0019PE", "DP03_0020E", "DP03_0020PE", "DP03_0021E", "DP03_0021PE", "DP03_0022E", "DP03_0022PE", "DP03_0023E", "DP03_0023PE", "DP03_0024E", "DP03_0024PE", "DP03_0025E", "DP03_0025PE"], 
+    "EDUCATIONAL ATTAINMENT": ["DP02_0059E", "DP02_0059PE", "DP02_0060E", "DP02_0060PE", "DP02_0061E", "DP02_0061PE", "DP02_0062E", "DP02_0062PE", "DP02_0063E", "DP02_0063PE", "DP02_0064E", "DP02_0064PE", "DP02_0065E", "DP02_0065PE", "DP02_0066E", "DP02_0066PE", "DP02_0067E", "DP02_0067PE", "DP02_0068E", "DP02_0068PE"], 
+    "EMPLOYMENT STATUS ": ["DP03_0001E", "DP03_0001PE", "DP03_0002E", "DP03_0002PE", "DP03_0003E", "DP03_0003PE", "DP03_0004E", "DP03_0004PE", "DP03_0005E", "DP03_0005PE", "DP03_0006E", "DP03_0006PE", "DP03_0007E", "DP03_0007PE", "DP03_0008E", "DP03_0008PE", "DP03_0009E", "DP03_0009PE", "DP03_0010E", "DP03_0010PE", "DP03_0011E", "DP03_0011PE", "DP03_0012E", "DP03_0012PE", "DP03_0013E", "DP03_0013PE", "DP03_0014E", "DP03_0014PE", "DP03_0015E", "DP03_0015PE", "DP03_0016E", "DP03_0016PE", "DP03_0017E", "DP03_0017PE"], 
+    "HOUSE HEATING FUEL ": ["DP04_0062E", "DP04_0062PE", "DP04_0063E", "DP04_0063PE", "DP04_0064E", "DP04_0064PE", "DP04_0065E", "DP04_0065PE", "DP04_0066E", "DP04_0066PE", "DP04_0067E", "DP04_0067PE", "DP04_0068E", "DP04_0068PE", "DP04_0069E", "DP04_0069PE", "DP04_0070E", "DP04_0070PE", "DP04_0071E", "DP04_0071PE"], 
+    "INCOME AND BENEFITS": ["DP03_0051E", "DP03_0051PE", "DP03_0052E", "DP03_0052PE", "DP03_0053E", "DP03_0053PE", "DP03_0054E", "DP03_0054PE", "DP03_0055E", "DP03_0055PE", "DP03_0056E", "DP03_0056PE", "DP03_0057E", "DP03_0057PE", "DP03_0058E", "DP03_0058PE", "DP03_0059E", "DP03_0059PE", "DP03_0060E", "DP03_0060PE", "DP03_0061E", "DP03_0061PE", "DP03_0062E", "DP03_0062PE", "DP03_0063E", "DP03_0063PE"], 
+    "INDUSTRY": ["DP03_0032E", "DP03_0032PE", "DP03_0033E", "DP03_0033PE", "DP03_0034E", "DP03_0034PE", "DP03_0035E", "DP03_0035PE", "DP03_0036E", "DP03_0036PE", "DP03_0037E", "DP03_0037PE", "DP03_0038E", "DP03_0038PE", "DP03_0039E", "DP03_0039PE", "DP03_0040E", "DP03_0040PE", "DP03_0041E", "DP03_0041PE", "DP03_0042E", "DP03_0042PE", "DP03_0043E", "DP03_0043PE", "DP03_0044E", "DP03_0044PE", "DP03_0045E", "DP03_0045PE"], 
+    "OCCUPATION": ["DP03_0026E", "DP03_0026PE", "DP03_0027E", "DP03_0027PE", "DP03_0028E", "DP03_0028PE", "DP03_0029E", "DP03_0029PE", "DP03_0030E", "DP03_0030PE", "DP03_0031E"], 
+    "PERCENTAGE under POVERTY LEVEL": ["DP03_0128PE"], 
+    "RACE": ["DP05_0033E", "DP05_0033PE", "DP05_0036E", "DP05_0036PE", "DP05_0037E", "DP05_0037PE", "DP05_0038E", "DP05_0038PE", "DP05_0039E", "DP05_0039PE", "DP05_0044E", "DP05_0044PE", "DP05_0052E", "DP05_0052PE", "DP05_0057E", "DP05_0057PE", "DP05_0058E", "DP05_0058PE"]
+}
